+++ Localization/deDE.lua
@@ -11,20 +11,24 @@ L["TAB"] = tab;
 -- Load strings into the localization that are already localized by Blizzard
 addon.GlobalStrings.Load(L);
 
-L["Default value"] = "Vorgabewert (Standard)";
-L["Checked"] = "Aktivert";
-L["Unchecked"] = "Nicht aktiviert";
-L["Build"] = "Version";
-L["Author"] = "Autor";
-L["Discord"] = "Discord";
-L["Discord Desc"] = "Öffnet ein Popup-Fenster mit einem Link zum {serverName} Discord-Server. Hier können Sie Kommentare, Berichte, Bemerkungen, Ideen und alles andere posten.";
-L["CurseForge"] = "CurseForge";
-L["CurseForge Desc"] = "Öffnet ein Popup-Fenster mit einem Link zur Seite {addonName} {curseForge}.";
-L["Wago"] = "Wago";
-L["Wago Desc"] = "Öffnet ein Popup-Fenster mit einem Link zur Seite {addonName} {wago}.";
-L["WoWInterface"] = "WoWInterface";
-L["WoWInterface Desc"] = "Öffnet ein Popup-Fenster mit einem Link zur Seite {addonName} {woWInterface}.";
-L["Show minimap icon"] = "Zeige Minimap Icon";
-L["Show minimap icon Desc"] = "Zeige / Verstecke das Minimap Icon.";
-L["Right click"] = "Rechts-Klick";
-L["Icon Right click"] = "für die Optionen.";
+-- [[ https://legacy.curseforge.com/wow/addons/krowi-merchant-frame-extended/localization ]] --
+-- [[ Everything after this line is automatically generated from CurseForge and is not meant for manual edit - SOURCETOKEN - AUTOGENTOKEN ]] --
+
+L["Author"] = "Autor"
+L["Build"] = "Version"
+L["Checked"] = "Aktivert"
+L["CurseForge"] = true
+L["CurseForge Desc"] = "Öffnet ein Popup-Fenster mit einem Link zur Seite {addonName} {curseForge}."
+L["Default value"] = "Vorgabewert (Standard)"
+L["Discord"] = true
+L["Discord Desc"] = "Öffnet ein Popup-Fenster mit einem Link zum {serverName} Discord-Server. Hier können Sie Kommentare, Berichte, Bemerkungen, Ideen und alles andere posten."
+L["Icon Right click"] = "für die Optionen."
+L["Right click"] = "Rechts-Klick"
+L["Show minimap icon"] = "Zeige Minimap Icon"
+L["Show minimap icon Desc"] = "Zeige / Verstecke das Minimap Icon."
+L["Unchecked"] = "Nicht aktiviert"
+L["Wago"] = true
+L["Wago Desc"] = "Öffnet ein Popup-Fenster mit einem Link zur Seite {addonName} {wago}."
+L["WoWInterface"] = true
+L["WoWInterface Desc"] = "Öffnet ein Popup-Fenster mit einem Link zur Seite {addonName} {woWInterface}."
+