+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 2.0 - 2023-08-09
+### Added
+- Now uses the Addon compartment (minimap button is still available for those who use it)
+- Introduced Profiles
+
+### Changed
+- Lots of backend code is now shared between all my addons
+
 ## 1.0 - 2023-07-22
 ### Added
 - First release